+++ src/datasource/LocationDatasource.ts
@@ -1,4 +1,5 @@
 import { BASE_URL, httpClient } from "../config/develop-config";
+import { SubdistrictEntity } from "../entities/LocationEntities";
 
 export class LocationDatasource {
   static getProvince(): Promise<any> {
@@ -12,9 +13,8 @@ export class LocationDatasource {
       });
   }
   static getDistrict(id: number): Promise<any> {
-
     return httpClient
-      .get(BASE_URL + "/location/district/"+id)
+      .get(BASE_URL + "/location/district/" + id)
       .then((response) => {
         return response.data;
       })
@@ -25,6 +25,19 @@ export class LocationDatasource {
   static getSubdistrict(id: number): Promise<any> {
     return httpClient
       .get(BASE_URL + "/location/sub-district/" + id)
+      .then((response) => {
+        return response.data;
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }
+  static getSearchLocation(text?: string): Promise<SubdistrictEntity[]> {
+    const param = {
+      text: text
+    }
+    return httpClient
+      .post(BASE_URL + "/location/sub-district" , param)
       .then((response) => {
         return response.data;
       })